Fix SonarQube issues from pages 2-6: 1,540 automated fixes (public modifiers, lambda to method refs)

+++ modules/trishul-money/src/main/java/io/trishul/money/serialize/MoneySerializer.java
@@ -19,7 +19,7 @@ public class MoneySerializer extends JsonSerializer<Money> {
       gen.writeStartObject();
       MoneyDto dto = MoneyMapper.INSTANCE.toDto(value);
       gen.writeStringField(MoneyDto.ATTR_CURRENCY, dto.getCurrency());
-      gen.writeNumberField(MoneyDto.ATTR_AMOUNT, (BigDecimal) dto.getAmount());
+      gen.writeNumberField(MoneyDto.ATTR_AMOUNT, dto.getAmount());
       gen.writeEndObject();
     }
   }
